requirements.txt
wheel
spacy==3.2.0
google-cloud-translate==2.0.1
python-dotenv==0.19.*
streamlit==1.3.*
pandas==1.3.*
iso639==0.1.*
gtts==2.2.*
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.2.0/de_core_news_sm-3.2.0-py3-none-any.whl
https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.2.0/pt_core_news_sm-3.2.0-py3-none-any.whl
https://github.com/explosion/spacy-models/releases/download/zh_core_web_sm-3.2.0/zh_core_web_sm-3.2.0-py3-none-any.whl
https://github.com/explosion/spacy-models/releases/download/ru_core_news_sm-3.2.0/ru_core_news_sm-3.2.0-py3-none-any.whl
https://github.com/explosion/spacy-models/releases/download/ro_core_news_sm-3.2.0/ro_core_news_sm-3.2.0-py3-none-any.whl
https://github.com/explosion/spacy-models/releases/download/pl_core_news_sm-3.2.0/pl_core_news_sm-3.2.0-py3-none-any.whl
https://github.com/explosion/spacy-models/releases/download/nl_core_news_sm-3.2.0/nl_core_news_sm-3.2.0-py3-none-any.whl
https://github.com/explosion/spacy-models/releases/download/nb_core_news_sm-3.2.0/nb_core_news_sm-3.2.0-py3-none-any.whl
https://github.com/explosion/spacy-models/releases/download/mk_core_news_sm-3.2.0/mk_core_news_sm-3.2.0-py3-none-any.whl
https://github.com/explosion/spacy-models/releases/download/lt_core_news_sm-3.2.0/lt_core_news_sm-3.2.0-py3-none-any.whl
https://github.com/explosion/spacy-models/releases/download/ja_core_news_sm-3.2.0/ja_core_news_sm-3.2.0-py3-none-any.whl
https://github.com/explosion/spacy-models/releases/download/it_core_news_sm-3.2.0/it_core_news_sm-3.2.0-py3-none-any.whl
https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.2.0/fr_core_news_sm-3.2.0-py3-none-any.whl
https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.2.0/es_core_news_sm-3.2.0-py3-none-any.whl
https://github.com/explosion/spacy-models/releases/download/el_core_news_sm-3.2.0/el_core_news_sm-3.2.0-py3-none-any.whl
https://github.com/explosion/spacy-models/releases/download/da_core_news_sm-3.2.0/da_core_news_sm-3.2.0-py3-none-any.whl
https://github.com/explosion/spacy-models/releases/download/ca_core_news_sm-3.2.0/ca_core_news_sm-3.2.0-py3-none-any.whl
